- ApiClientFactory impl: thread-safe ApiClient usage - refactor http Client init into HttpClientFactory - bump java to 1.8
